feat: move logic out for different func
