+++ magefiles/test-e2e-noncluster.go
@@ -1,0 +1,25 @@
+// Copyright 2026 colonel-byte
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//	http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+//go:build mage
+// +build mage
+
+package main
+
+// EndToEndNonCluster runs the group that needs no cluster: the misc and package command
+// groups. -short additionally skips the example packages, so this finishes in seconds.
+// Mirrors the e2e-noncluster CI job.
+func (Test) EndToEndNonCluster() error {
+	return runE2E("30m", "github.com/colonel-byte/cargoship/src/test/e2e/noncluster/...", "-short")
+}
